+++ src/Akeneo/Category/back/Domain/ValueObject/Code.php
@@ -14,7 +14,8 @@ class Code
 {
     public function __construct(private string $code)
     {
-        Assert::notEmpty($code);
+        Assert::notNull($code);
+        Assert::notEq($code, '');
     }
 
     public function __toString(): string