+++ CHANGELOG.md
@@ -109,6 +109,7 @@
 - PIM:10741: Fix diff indexation of product models
 - PIM:10743: Fix HTTP 500 on measurement PATCH without unit
 - PIM:10744: Fix product import with a quantified association column is missing
+- PIM-10750: Fix category code validation to allow '0'
 
 ## Improvements